220828 장바구니,바로구매 에서 결제나누기 완료

+++ src/main/java/controller/OrderController.java
@@ -1,14 +1,14 @@
 package controller;
 
-import java.util.Enumeration;
 import java.util.List;
 
 import javax.annotation.Resource;
-import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -33,15 +33,17 @@ public class OrderController {
 	@Resource(name = "loginUserBean")
 	private UserBean loginUserBean;
 
-	@PostMapping("/order_pro")
+	@PostMapping("/main")
 	public String order_pro(@ModelAttribute("orderInfoBean") OrderBean orderInfoBean,
 							@ModelAttribute("basketOrderBean") OrderBean basketOrderBean,
 							@RequestParam("goods_idx") String goods_idx,
+//							@RequestParam("user_idx") String user_idx,
 							@RequestParam(value = "order_basket", required = false) String order_basket, 
 							@RequestParam(value = "topmenu_name", required = false) String topmenu_name,
 							Model model) {
 		
 		model.addAttribute("goods_idx", goods_idx);
+//		model.addAttribute("user_idx", user_idx);
 		model.addAttribute("topmenu_name", topmenu_name);
 		model.addAttribute("orderInfoBean", orderInfoBean);
 		model.addAttribute("basketOrderBean", basketOrderBean);
@@ -83,15 +85,32 @@ public class OrderController {
 	
 	@PostMapping("/pay_pro")
 	public String pay_pro(@ModelAttribute("payInfoBean") OrderBean payInfoBean,
-						   Model model) {
+						  Model model) {
 		
-		List<OrderBean> payInfo = orderService.addPayInfo(payInfoBean);
+		List<OrderBean> payInfo = orderService.getPayInfo(payInfoBean);	 // 주문정보 select
 		model.addAttribute("payInfo", payInfo);
-		model.addAttribute("order_pay_option", payInfoBean.getOrder_pay_option());
+//		model.addAttribute("order_pay_option", payInfoBean.getOrder_pay_option());
 		
-		orderService.deleteBasketInfo(loginUserBean.getUser_idx());
+		if(payInfoBean.getPayOK() != "1" ||
+				payInfoBean.getPayOK() != "0"	) {	// 장바구니 -> pay_success 라면
+			orderService.deleteBasketInfo(loginUserBean.getUser_idx());	// 결제완료 후 basket_table 목록 삭제
+		}
 		
+		orderService.modifyPayOptionInfo(payInfoBean);	// order_pay_option(결제수단) 추가
+		orderService.modifypayOKInfo(payInfoBean);		// payOK(결제완료) 0으로 변경
+		orderService.modifyPayInfo(payInfoBean);	// order_main -> pay_success '결제정보' 수정 반영
+		model.addAttribute("payInfoBean", payInfoBean);
 		return "order/pay_success";
+	}
+	
+	@GetMapping("/order_manage")
+	public String order_manage(OrderBean orderManageBean, Model model) {
+		
+		List<OrderBean> orderManageInfo = orderService.getOrderManageInfo(orderManageBean);
+		
+		model.addAttribute("orderManageInfo", orderManageInfo);
+		
+		return "admin/order_manage";
 	}
 
 }
